Added more aggressive locks to fix race condition
Race condition observed in igecs - sequence as follows (AFAICT - may also have been in consumption code)

1. Thread A invokes on_resolve to Promise A, noticing it isn't resolved
2. Thread B begins to resolve Promise A, flushing the callback list
3. Thread A adds callback to Promise A callback list

In this case, the callback added on Thread A will never be scheduled. Ever.

Unfortunately, as part of this, inline_execution_context had to be removed, since using it
caused deadlock in the library. Bummer.